Bot Weather
When the user writes start, the bot greets the user and writes to him: "Write the name of the city...". After the user has entered the city, the bot sends the current temperature and humidity in this city. If the user wrote not a real city, then the bot will answer: "Hmm... I don't know such a city. :( Try again..."